reduz a complexidade do componente
